requirements.txt
paho-mqtt
scikit-learn
pandas
numpy
streamlit
plotly
python-dotenv
schedule
prophet